requirements.txt
opencv-python
numpy
matplotlib
mysql-connector
pyserial
scipy
imutils
playsound
cmake
Pillow
DateTime
argparse
dlib==19.21.0
